+++ src/main/java/com/zcodegroup/ws/service/UserService.java
@@ -1,14 +1,13 @@
 package com.zcodegroup.ws.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.zcodegroup.ws.model.User;
 
 public interface UserService {
 	public Collection<User> findAll();
-	public User findOne(BigInteger id);
+	public User findOne(Long id);
 	public User create(User user);
 	public User update(User user);
-	public boolean delete(BigInteger id);
+	public boolean delete(Long id);
 }
